+++ src/components/trip-board.js
@@ -1,27 +1,12 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createBoardTripTemplate = () => {
   return `<ul class="trip-days">
   </ul>`;
 };
 
-export default class Board {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Board extends AbstractComponent {
   getTemplate() {
     return createBoardTripTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
