+++ src/components/trip-point.js
@@ -1,11 +1,20 @@
 // import {createOffersTemplate} from './trip-offers';
-import {formatTimeDuration, formatTime, createElement} from '../utils.js';
-import OffersComponent from './trip-offers.js';
+import {formatTimeDuration, formatTime} from '../utils/common.js';
+import AbstractComponent from './abstract-component.js';
+
+const createOffersTemplate = (offers) => {
+  return offers.map((it) => (
+    `<li class="event__offer">
+            <span class="event__offer-title">${it.name}</span>
+            &plus;
+            &euro;&nbsp;<span class="event__offer-price">${it.cost}</span>
+           </li>`
+  )).join(``);
+};
 
 const createTripPointTemplate = (tripPoint) => {
   const {type, destination, price, offers, startDate, endDate} = tripPoint;
   const duration = endDate - startDate;
-  const extraOffers = new OffersComponent(Array.from(offers)).getTemplate();
 
 
   let preposition = `to`;
@@ -35,7 +44,7 @@ const createTripPointTemplate = (tripPoint) => {
 
     <h4 class="visually-hidden">Offers:</h4>
     <ul class="event__selected-offers">
-    ${extraOffers}
+    ${createOffersTemplate(offers)}
     </ul>
 
     <button class="event__rollup-btn" type="button">
@@ -45,24 +54,18 @@ const createTripPointTemplate = (tripPoint) => {
 </li>`;
 };
 
-export default class TripPoint {
+export default class TripPoint extends AbstractComponent {
   constructor(tripPoint) {
+    super();
     this._tripPoint = tripPoint;
-    this._element = null;
   }
 
   getTemplate() {
     return createTripPointTemplate(this._tripPoint);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
+  setEditButtonHandler(handler) {
+    this.getElement().querySelector(`.event__rollup-btn`)
+    .addEventListener(`click`, handler);
   }
 }
